Update - Anthony 03/06/2020

+++ hris/app/Http/Controllers/BenefitController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use App\hris_benefits;
+use App\users;
 
 class BenefitController extends Controller
 {
@@ -61,9 +62,14 @@ class BenefitController extends Controller
 
     public function destroy(hris_benefits $benefit)
     {
-        $benefit->delete();
-        Log::channel('benefits')->info('Benefit id no.' .$benefit->id. ' has been deleted. Name: '. $benefit->name. '.');
-        return redirect('/hris/pages/recruitment/recruitmentSetup/benefits/index')->with('success','Benefit successfully deleted!');
+        $id = $_SESSION['sys_id'];
+        $upass = $this->decryptStr(users::find($id)->upass);
+        if ( $upass == request('upass') ) {
+            $benefit->delete();
+            return redirect('/hris/pages/recruitment/recruitmentSetup/benefits/index')->with('success','Benefit successfully deleted!');
+        } else {
+            return back()->withErrors(['Password does not match.']);
+        }
     }
 
     protected function validatedData()
@@ -71,5 +77,17 @@ class BenefitController extends Controller
         return request()->validate([
             'name' => 'required|max:100'
         ]);
+    }
+    // decrypt string
+    function decryptStr($str) {
+        $key = '4507';
+        $c = base64_decode($str);
+        $ivlen = openssl_cipher_iv_length($cipher="AES-128-CBC");
+        $iv = substr($c,0,$ivlen);
+        $hmac = substr($c,$ivlen,$sha2len=32);
+        $ciphertext_raw = substr($c,$ivlen+$sha2len);
+        $original_plaintext = openssl_decrypt($ciphertext_raw,$cipher,$key,$options=OPENSSL_RAW_DATA,$iv);
+        $calcmac = hash_hmac('sha256',$ciphertext_raw,$key,$as_binary=true);
+        if (hash_equals($hmac,$calcmac)) { return $original_plaintext; }
     }
 }
