+++ hris/app/Http/Controllers/JobFunctionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use App\hris_job_functions;
+use App\users;
 
 class JobFunctionController extends Controller
 {
@@ -56,8 +57,14 @@ class JobFunctionController extends Controller
 
     public function destroy(hris_job_functions $jobFunction)
     {
-        $jobFunction->delete();
-        return redirect('/hris/pages/recruitment/recruitmentSetup/jobFunctions/index')->with('success','Job function successfully deleted!');
+        $id = $_SESSION['sys_id'];
+        $upass = $this->decryptStr(users::find($id)->upass);
+        if ( $upass == request('upass') ) {
+            $jobFunction->delete();
+            return redirect('/hris/pages/recruitment/recruitmentSetup/jobFunctions/index')->with('success','Job function successfully deleted!');
+        } else {
+            return back()->withErrors(['Password does not match.']);
+        }
     }
 
     protected function validatedData() 
@@ -65,5 +72,17 @@ class JobFunctionController extends Controller
         return request()->validate([
             'name' => 'required|max:100'
         ]);
+    }
+    // decrypt string
+    function decryptStr($str) {
+        $key = '4507';
+        $c = base64_decode($str);
+        $ivlen = openssl_cipher_iv_length($cipher="AES-128-CBC");
+        $iv = substr($c,0,$ivlen);
+        $hmac = substr($c,$ivlen,$sha2len=32);
+        $ciphertext_raw = substr($c,$ivlen+$sha2len);
+        $original_plaintext = openssl_decrypt($ciphertext_raw,$cipher,$key,$options=OPENSSL_RAW_DATA,$iv);
+        $calcmac = hash_hmac('sha256',$ciphertext_raw,$key,$as_binary=true);
+        if (hash_equals($hmac,$calcmac)) { return $original_plaintext; }
     }
 }
