+++ hris/app/Http/Controllers/PayGradeController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use App\hris_pay_grades;
+use App\users;
 
 class PayGradeController extends Controller
 {
@@ -60,8 +61,14 @@ class PayGradeController extends Controller
 
     public function destroy(hris_pay_grades $payGrade)
     {
-        $payGrade->delete();
+        $id = $_SESSION['sys_id'];
+        $upass = $this->decryptStr(users::find($id)->upass);
+        if ( $upass == request('upass') ) {
+            $payGrade->delete();
             return redirect('/hris/pages/admin/jobDetails/payGrades/index')->with('success', 'Pay Grade successfully deleted!');
+        } else {
+            return back()->withErrors(['Password does not match.']);
+        }
     }
 
     protected function validatedData() {
@@ -73,6 +80,18 @@ class PayGradeController extends Controller
             'max_salary' => 'required'
         ]);
 
+    }
+    // decrypt string
+    function decryptStr($str) {
+        $key = '4507';
+        $c = base64_decode($str);
+        $ivlen = openssl_cipher_iv_length($cipher="AES-128-CBC");
+        $iv = substr($c,0,$ivlen);
+        $hmac = substr($c,$ivlen,$sha2len=32);
+        $ciphertext_raw = substr($c,$ivlen+$sha2len);
+        $original_plaintext = openssl_decrypt($ciphertext_raw,$cipher,$key,$options=OPENSSL_RAW_DATA,$iv);
+        $calcmac = hash_hmac('sha256',$ciphertext_raw,$key,$as_binary=true);
+        if (hash_equals($hmac,$calcmac)) { return $original_plaintext; }
     }
 
 }
