+++ hris/app/Http/Controllers/SkillController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Log;
 use App\hris_skills;
+use App\users;
 
 class SkillController extends Controller
 {
@@ -58,8 +59,14 @@ class SkillController extends Controller
 
     public function destroy(hris_skills $skill)
     {
-        $skill->delete();
-        return redirect('/hris/pages/admin/qualifications/skills/index')->with('success', 'Skill successfully deleted!');
+        $id = $_SESSION['sys_id'];
+        $upass = $this->decryptStr(users::find($id)->upass);
+        if ( $upass == request('upass') ) {
+            $skill->delete();
+            return redirect('/hris/pages/admin/qualifications/skills/index')->with('success', 'Skill successfully deleted!');
+        } else {
+            return back()->withErrors(['Password does not match.']);
+        }
     }
 
     protected function validatedData() 
@@ -68,5 +75,17 @@ class SkillController extends Controller
             'name' => 'required',
             'description' => 'required'
         ]);
+    }
+    // decrypt string
+    function decryptStr($str) {
+        $key = '4507';
+        $c = base64_decode($str);
+        $ivlen = openssl_cipher_iv_length($cipher="AES-128-CBC");
+        $iv = substr($c,0,$ivlen);
+        $hmac = substr($c,$ivlen,$sha2len=32);
+        $ciphertext_raw = substr($c,$ivlen+$sha2len);
+        $original_plaintext = openssl_decrypt($ciphertext_raw,$cipher,$key,$options=OPENSSL_RAW_DATA,$iv);
+        $calcmac = hash_hmac('sha256',$ciphertext_raw,$key,$as_binary=true);
+        if (hash_equals($hmac,$calcmac)) { return $original_plaintext; }
     }
 }
