+++ hris/app/Http/Controllers/TimeProjectController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\hris_projects;
 use App\hris_time_projects;
+use App\users;
 
 class TimeProjectController extends Controller
 {
@@ -60,8 +61,14 @@ class TimeProjectController extends Controller
 
     public function destroy(hris_time_projects $timeProject)
     {
-        $timeProject->delete();
-        return redirect('/hris/pages/time/timeProjects/index')->with('success', 'Project successfully deleted!');
+        $id = $_SESSION['sys_id'];
+        $upass = $this->decryptStr(users::find($id)->upass);
+        if ( $upass == request('upass') ) {
+            $timeProject->delete();
+            return redirect('/hris/pages/time/timeProjects/index')->with('success', 'Project successfully deleted!');
+        } else {
+            return back()->withErrors(['Password does not match.']);
+        }
     }
 
     protected function validatedData()
@@ -69,6 +76,18 @@ class TimeProjectController extends Controller
         return request()->validate([
             'project' => 'required'
         ]);
+    }
+    // decrypt string
+    function decryptStr($str) {
+        $key = '4507';
+        $c = base64_decode($str);
+        $ivlen = openssl_cipher_iv_length($cipher="AES-128-CBC");
+        $iv = substr($c,0,$ivlen);
+        $hmac = substr($c,$ivlen,$sha2len=32);
+        $ciphertext_raw = substr($c,$ivlen+$sha2len);
+        $original_plaintext = openssl_decrypt($ciphertext_raw,$cipher,$key,$options=OPENSSL_RAW_DATA,$iv);
+        $calcmac = hash_hmac('sha256',$ciphertext_raw,$key,$as_binary=true);
+        if (hash_equals($hmac,$calcmac)) { return $original_plaintext; }
     }
 
 }
